requirements.txt
beautifulsoup4==4.7.1
bs4==0.0.1
cycler==0.10.0
kiwisolver==1.1.0
matplotlib==3.0.3
nltk==3.4.1
numpy==1.16.3
pandas==0.24.2
pkg-resources==0.0.0
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
simplejson==3.16.0
six==1.12.0
soupsieve==1.9.1
textblob==0.15.3
